Update ci.yml (#65)
Don't delete removed files when deploying

Our current build attaches content hashes to some filenames, which change when the files are updated. But we've run into problems where an old version of an un-hashed file (e.g. index.html) links to a deleted filename that wasn't itself hashed, causing the viewer to fail to load (this happend with both cloudfront and browser caches). There's probably an elegant solution if you understand http headers and web caching, but in the meantime we'll just leave the old versions sitting around.